# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LeadChain API",
        "description": "API REST de LeadChain para la gestión comercial de clientes, edificios y visitas. Autenticación mediante JWT (Bearer token). Incluye soporte para coordenadas geográficas PostGIS.",
        "contact": {
            "name": "Alberto Romero Pino",
            "email": "albertoromeropino2004@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener todos los usuarios",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un nuevo usuario",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "d3a14968f8b7a072620c75c7c78ed6f8",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario completo",
                "operationId": "e29fc231c7937ab09483d35e6e3d1214",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "e0f9a385a1d6788b99e519bd98cf7239",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuario eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar parcialmente un usuario",
                "operationId": "2eea0b6e40fa6e600a841277b6e2ff01",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado por rol"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/comerciales-a-cargo": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener comerciales a cargo del usuario actual con visitas consolidadas",
                "description": "Obtener comerciales del usuario actual con sus visitas, clientes y zonas\nConsolida en una sola petición toda la información necesaria para ComercialesPage",
                "operationId": "4cdcaa35448fac6f48005dde43585acf",
                "responses": {
                    "200": {
                        "description": "Comerciales con visitas, clientes y zonas consolidados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "comerciales": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        },
                                        "visitas": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/VisitaResource"
                                            }
                                        },
                                        "clientes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ClienteResource"
                                            }
                                        },
                                        "zonas": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ZonaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (solo admin)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "GeoPoint": {
                "required": [
                    "lat",
                    "lng"
                ],
                "properties": {
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "example": 37.9
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "example": -4.8
                    }
                },
                "type": "object"
            },
...
            "ClienteResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Antonio"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Perez Garcia"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "612345678"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "antonio@ejemplo.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Alejandro"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Garcia Martinez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ale.garcia@email.com"
                    },
                    "rol": {
                        "type": "string",
                        "example": "comercial"
                    },
                    "id_responsable": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "id_zona": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "VisitaResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_usuario": {
                        "type": "integer",
                        "example": 2
                    },
                    "id_cliente": {
                        "type": "integer",
                        "example": 1
                    },
                    "fecha_hora": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-03-15 10:30:00"
                    },
                    "id_estado": {
                        "type": "integer",
                        "example": 1
                    },
                    "observaciones": {
                        "type": "string",
                        "example": "Primera visita al cliente",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ZonaResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Zona Norte"
                    },
                    "area": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GeoPoint"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}